BlackoilModel: move properties to separate header
